Log entry page seems to be working with currently supported fields. More tests to come.
